+++ CHANGELOG.md
@@ -3,6 +3,8 @@ CHANGELOG
 
 Upcoming
 --------
+- Instrs-logging now support qubit IDs and qubit states for all qubits involved in the operation.
+- Instrs-logging now support what to specify what qubits have at some point interacted in the simulaton.
 
 2020-08-27 (0.0.8)
 ------------------